web_api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
itsdangerous>=2.2.0
jinja2>=3.1.4
python-dotenv>=1.0.0
PyYAML>=6.0.1
python-multipart>=0.0.9
aiosqlite>=0.19.0
discord.py>=2.3.0
